Refactor asset workspace serializers
